requirements.txt
albucore==0.0.24
albumentations==2.0.8
annotated-types==0.7.0
certifi==2025.4.26
charset-normalizer==3.4.2
filelock==3.18.0
fsspec==2025.5.1
hf-xet==1.1.2
huggingface-hub==0.32.3
idna==3.10
joblib==1.5.1
numpy==1.26.4
opencv-python-headless==4.11.0.86
packaging==25.0
pandas==2.2.3
pillow==11.2.1
pydantic==2.11.5
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
simsimd==6.2.1
six==1.17.0
stringzilla==3.12.5
threadpoolctl==3.6.0
timm==1.0.15
tokenizers==0.13.3
torch==1.13.1+cu116
torchaudio==0.13.1+cu116
torchvision==0.14.1+cu116
tqdm==4.67.1
transformers==4.27.4
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
